+++ chevah/compat/tests/normal/test_filesystem.py
@@ -10,6 +10,7 @@ import stat
 from mock import patch
 
 from chevah.compat import DefaultAvatar, LocalFilesystem
+from chevah.compat.exceptions import CompatError
 from chevah.compat.interfaces import ILocalFilesystem
 from chevah.compat.testing import CompatTestCase, conditionals, manufacture
 
@@ -640,6 +641,41 @@ class TestDefaultFilesystem(FilesystemTestCase):
         self.assertTrue(stat.S_ISREG(status.st_mode))
         self.assertFalse(stat.S_ISDIR(status.st_mode))
         self.assertFalse(stat.S_ISLNK(status.st_mode))
+
+    #@conditionals.onOSFamily('posix')
+    def test_checkChildPath_unix(self):
+        """
+        Will raise an error if child is outside of root, or do nothing if
+        child is is root.
+        """
+        self.filesystem._checkChildPath(u'/root/path', '/root/path')
+        self.filesystem._checkChildPath(u'/root/path/', '/root/path')
+        self.filesystem._checkChildPath(u'/root/path', '/root/path/')
+        self.filesystem._checkChildPath(u'/root/path', '/root/path/../path/')
+
+        with self.assertRaises(CompatError):
+            self.filesystem._checkChildPath(u'/root/path', '/')
+
+        with self.assertRaises(CompatError):
+            self.filesystem._checkChildPath(u'/root/path', '/root/path/..')
+
+    @conditionals.onOSFamily('nt')
+    def test_checkChildPath_nt(self):
+        """
+        See: test_checkChildPath_unix
+        """
+        self.filesystem._checkChildPath(u'c:\\root\\path', 'c:\\root\\path')
+        self.filesystem._checkChildPath(u'c:\\root\\path\\', 'c:\\root\\path')
+        self.filesystem._checkChildPath(u'c:\\root\\path', 'c:\\root\\path\\')
+        self.filesystem._checkChildPath(
+            u'c:\\root\\path', 'c:\\root\\path\\..\\path')
+
+        with self.assertRaises(CompatError):
+            self.filesystem._checkChildPath(u'c:\\root\\path', 'c:\\')
+
+        with self.assertRaises(CompatError):
+            self.filesystem._checkChildPath(
+                u'c:\\root\\path', 'c:\\root\\path\\..')
 
 
 class TestPosixFilesystem(CompatTestCase):
@@ -1500,6 +1536,28 @@ class TestLocalFilesystemLocked(CompatTestCase):
             root_path + separator + name + separator + child + separator)
         self.assertEqual([name, child], result)
 
+    @conditionals.onOSFamily('posix')
+    def test_getSegmentsFromRealPath_outside_home_unix(self):
+        """
+        Raise CompatError when path is outside of home.
+        """
+        with self.assertRaises(CompatError):
+            self.locked_filesystem.getSegmentsFromRealPath('/outside/path')
+
+        with self.assertRaises(CompatError):
+            self.locked_filesystem.getSegmentsFromRealPath('../../outside')
+
+    @conditionals.onOSFamily('nt')
+    def test_getSegmentsFromRealPath_outside_home_nt(self):
+        """
+        Raise CompatError when path is outside of home.
+        """
+        with self.assertRaises(CompatError):
+            self.locked_filesystem.getSegmentsFromRealPath('c:\\outside\\home')
+
+        with self.assertRaises(CompatError):
+            self.locked_filesystem.getSegmentsFromRealPath('..\\..\\outside')
+
     def test_exists_false(self):
         """
         exists will return `False` if file or folder does not exists.
@@ -1538,6 +1596,43 @@ class TestLocalFilesystemLocked(CompatTestCase):
             self.assertTrue(self.locked_filesystem.exists(segments))
         finally:
             self.locked_filesystem.deleteFolder(segments)
+
+    @conditionals.onCapability('symbolic_link', True)
+    def test_exists_broken_link(self):
+        """
+        Will return false when link target does not exists.
+        """
+        name = manufacture.makeFilename(suffix='-link')
+        segments = self.locked_filesystem.home_segments[:]
+        segments.append(name)
+        # Register for cleanup.
+        self.test_segments = manufacture.fs.temp_segments
+        self.test_segments.append(name)
+        self.locked_filesystem.makeLink(
+            target_segments=['z', 'no-such', 'target'],
+            link_segments=segments,
+            )
+
+        self.assertFalse(self.locked_filesystem.exists(segments))
+        # Link still exists.
+        self.assertTrue(self.locked_filesystem.isLink(segments))
+
+    @conditionals.onCapability('symbolic_link', True)
+    def test_readLink_outside_home(self):
+        """
+        Raise an error when target is outside of locked folder.
+        """
+        _, self.test_segments = manufacture.fs.makePathInTemp()
+        link_segments = [self.test_segments[-1]]
+        manufacture.fs.makeLink(
+            target_segments=['z', 'no', 'such'],
+            link_segments=self.test_segments,
+            )
+        # Make sure link was created.
+        self.assertTrue(self.locked_filesystem.isLink(link_segments))
+
+        with self.assertRaises(CompatError):
+            self.locked_filesystem.readLink(link_segments)
 
     def test_touch(self):
         """