+++ release-notes.rst
@@ -1,5 +1,12 @@
-Release notes for chevah.empirical
-==================================
+Release notes for chevah.compat
+===============================
+
+
+0.18.0 - 24/03/2014
+-------------------
+
+* Raise CompatError in getSegmentsFromRealPath if path is outside of home
+  folder.
 
 
 0.17.1 - 20/03/2014